requirements.txt
django>=4.2,<5.0
django-rest-framework>=3.14.0
psycopg2-binary>=2.9.5  # For PostgreSQL
gunicorn>=20.1.0  # WSGI server
python-dotenv>=0.21.0
whitenoise>=6.3.0  # For static files
dj-database-url>=1.2.0  # For database URL configuration